Move main function to own dedicated file
- file 'SW_Main.c' is now only used for SOILWAT2 standalone executable
- all other global definitions are in SW_Main_lib.c
